f-mvp: restructuring of files and folders
