--- src/App.tsx
@@ -1,12 +1,0 @@
-import * as React from "react";
-import "./App.scss";
-
-function App() {
-  return (
-    <div className="App">
-      <h1>App</h1>
-    </div>
-  );
-}
-
-export default App;
