+++ src/app/App.tsx
@@ -1,0 +1,12 @@
+import * as React from "react";
+import "./App.scss";
+
+import { Home } from "./pages/Home";
+
+export const App = () => {
+  return (
+    <div className="app">
+      <Home />
+    </div>
+  );
+};
