+++ src/app/components/Footer/Footer.scss
@@ -1,0 +1,2 @@
+// SCSS base
+@import "../../App.scss";
