+++ src/app/components/Footer/Footer.tsx
@@ -1,0 +1,10 @@
+import * as React from "react";
+import "./Footer.scss";
+
+export const Footer = (): JSX.Element => {
+  return (
+    <div>
+      <h1>Footer</h1>
+    </div>
+  );
+};
