+++ src/app/components/Header/Header.scss
@@ -1,0 +1,2 @@
+// SCSS base
+@import "../../App.scss";
