+++ src/app/components/Header/Header.tsx
@@ -1,0 +1,25 @@
+import * as React from "react";
+import "./Header.scss";
+
+import logo from "../../../assets/logo.svg";
+
+export const Header = (): JSX.Element => {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+        <p>
+          Edit <code>src/App.tsx</code> and save to reload.
+        </p>
+        <a
+          className="App-link"
+          href="https://reactjs.org"
+          target="_blank"
+          rel="noopener noreferrer"
+        >
+          Learn React
+        </a>
+      </header>
+    </div>
+  );
+};
