+++ src/app/pages/Home.tsx
@@ -1,0 +1,13 @@
+import * as React from "react";
+
+import { Header } from "../components/Header/Header";
+import { Footer } from "../components/Footer/Footer";
+
+export const Home = (): JSX.Element => {
+  return (
+    <div>
+      <Header />
+      <Footer />
+    </div>
+  );
+};
