--- src/components/Footer/Footer.tsx
@@ -1,13 +1,0 @@
-import * as React from "react";
-
-import "./Footer.scss";
-
-export class Footer extends React.Component {
-  render() {
-    return (
-      <div>
-        <h1>Footer</h1>
-      </div>
-    );
-  }
-}
