--- src/components/Header/Header.tsx
@@ -1,13 +1,0 @@
-import * as React from "react";
-
-import "./Header.scss";
-
-export class Header extends React.Component {
-  render() {
-    return (
-      <div>
-        <h1>Header</h1>
-      </div>
-    );
-  }
-}
