--- src/pages/Home.tsx
@@ -1,11 +1,0 @@
-import * as React from "react";
-
-export class Home extends React.Component {
-  render() {
-    return (
-      <div>
-        <h1>Home</h1>
-      </div>
-    );
-  }
-}
