+++ tsconfig.json
@@ -20,7 +20,7 @@
     "jsx": "react"
   },
   "include": [
-    "src/**/*"
+    "src/**/*", "src/types/custom.d.ts"
   ],
   "exclude": [
     "node_modules",
